+++ tiledb/sm/buffer/const_buffer.cc
@@ -33,6 +33,8 @@
 #include "tiledb/sm/buffer/const_buffer.h"
 #include "tiledb/sm/buffer/buffer.h"
 
+#include <iostream>
+
 namespace tiledb {
 namespace sm {