+++ tiledb/sm/c_api/tiledb.cc
@@ -3236,6 +3236,21 @@ int32_t tiledb_array_consolidate_with_key(
   return TILEDB_OK;
 }
 
+int32_t tiledb_array_vacuum(
+    tiledb_ctx_t* ctx, const char* array_uri, tiledb_config_t* config) {
+  // Sanity checks
+  if (sanity_check(ctx) == TILEDB_ERR)
+    return TILEDB_ERR;
+
+  if (SAVE_ERROR_CATCH(
+          ctx,
+          ctx->ctx_->storage_manager()->array_vacuum(
+              array_uri, (config == nullptr) ? nullptr : config->config_)))
+    return TILEDB_ERR;
+
+  return TILEDB_OK;
+}
+
 int32_t tiledb_array_get_non_empty_domain(
     tiledb_ctx_t* ctx, tiledb_array_t* array, void* domain, int32_t* is_empty) {
   if (sanity_check(ctx) == TILEDB_ERR || sanity_check(ctx, array) == TILEDB_ERR)